dfs: refactor: Update barometric pressure setting calculation
This commit updates the calculation of the barometric pressure setting in the BDS 4.0 decoder. The previous calculation used a range of 0 to 410, but it has been updated to use a range of 800 to 1210. This change ensures that the barometric pressure setting is within the correct range and improves the accuracy of the decoding process.